UI redesign, event/club/student fixes, visibility toggle, circular JSON fix

+++ anvay/anvay/src/main/java/com/cts/mfrp/anvay/controller/AchievementController.java
@@ -10,12 +10,12 @@ import org.springframework.web.bind.annotation.*;
 public class AchievementController {
 
     @GetMapping("/{achievementId}")
-    public ResponseEntity<Void> getAchievement(@PathVariable Integer achievementId) {
+    public ResponseEntity<Void> getAchievement(@PathVariable Long achievementId) {
         return ResponseEntity.ok().build();
     }
 
     @GetMapping("/user/{userId}")
-    public ResponseEntity<Void> getUserAchievements(@PathVariable Integer userId) {
+    public ResponseEntity<Void> getUserAchievements(@PathVariable Long userId) {
         return ResponseEntity.ok().build();
     }
 }
