+++ anvay/anvay/src/main/java/com/cts/mfrp/anvay/dto/RegisterStudentRequest.java
@@ -16,5 +16,5 @@ public class RegisterStudentRequest {
     @NotBlank(message = "Password is required")
     private String password;
 
-    private Integer institutionId;
+    private Long institutionId;
 }
